# backend/swagger.yaml
openapi: 3.0.0
info:
  title: My API
  version: 1.0.0
  description: API для управления пользователями и товарами
  contact:
    name: Support
    email: support@example.com

servers:
  - url: http://localhost:3000
    description: Development server

tags:
  - name: Auth
    description: Аутентификация и регистрация пользователей
  - name: Products
    description: Управление товарами

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT токен, полученный при логине (accessToken)"

  schemas:
    # Схема пользователя (ответ)
    User:
      type: object
      properties:
        id:
          type: string
          example: "abc123"
          description: Уникальный ID пользователя
        username:
          type: string
          example: "user@example.com"
          description: Email пользователя (используется как username)
      example:
        id: "abc123"
        username: "user@example.com"

    # Схема пользователя (регистрация, полная)
    UserFull:
      type: object
      properties:
        id:
          type: string
          example: "abc123"
        username:
          type: string
          example: "user@example.com"
        first_name:
          type: string
          example: "Иван"
        last_name:
          type: string
          example: "Иванов"
        hashedPassword:
          type: string
          example: "$2b$10$..."
          description: Хешированный пароль (не возвращается в ответе)

    # Схема для регистрации (входные данные)
    RegisterInput:
      type: object
      required:
        - email
        - first_name
        - last_name
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
          description: Email пользователя (будет использован как username)
        first_name:
          type: string
          example: "Иван"
        last_name:
          type: string
          example: "Иванов"
        password:
          type: string
          format: password
          example: "qwerty123"
      example:
        email: "user@example.com"
        first_name: "Иван"
        last_name: "Иванов"
        password: "qwerty123"

    # Схема для входа
    LoginInput:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "user@example.com"
          description: Имя пользователя (email)
        password:
          type: string
          format: password
          example: "qwerty123"
      example:
        username: "user@example.com"
        password: "qwerty123"

    # Схема успешного входа (с токенами)
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (используется для Bearer авторизации)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: JWT refresh token (используется для обновления пары токенов)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    # Схема для запроса обновления токенов
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Действующий refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      example:
        refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    # Схема ответа при обновлении токенов
    RefreshTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: Новый access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: Новый refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    # Схема товара
    Product:
      type: object
      properties:
        id:
          type: string
          example: "xyz789"
        title:
          type: string
          example: "Ноутбук"
        category:
          type: string
          example: "Электроника"
        description:
          type: string
          example: "Мощный ноутбук для работы и игр"
        price:
          type: number
          example: 50000
      example:
        id: "xyz789"
        title: "Ноутбук"
        category: "Электроника"
        description: "Мощный ноутбук для работы и игр"
        price: 50000

    # Схема для создания/обновления товара (входные данные)
    ProductInput:
      type: object
      required:
        - title
        - category
        - description
        - price
      properties:
        title:
          type: string
          example: "Ноутбук"
        category:
          type: string
          example: "Электроника"
        description:
          type: string
          example: "Мощный ноутбук для работы и игр"
        price:
          type: number
          example: 50000
      example:
        title: "Ноутбук"
        category: "Электроника"
        description: "Мощный ноутбук для работы и игр"
        price: 50000

    # Схема ошибки
    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
      example:
        error: "Not found"

paths:
  # --- АУТЕНТИФИКАЦИЯ ---
  /auth/register:
    post:
      tags:
        - Auth
      summary: Регистрация нового пользователя
      description: Создает нового пользователя с переданными данными
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        201:
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserFull"
        400:
          description: Не все обязательные поля заполнены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/login:
    post:
      tags:
        - Auth
      summary: Вход в систему
      description: Аутентификация пользователя по username и password. Возвращает access и refresh токены.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        200:
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        400:
          description: Не указаны username или password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Неверный пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Получить информацию о текущем пользователе
      description: Возвращает данные аутентифицированного пользователя (без пароля).
      security:
        - bearerAuth: []
      responses:
        200:
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        401:
          description: Не авторизован (отсутствует или неверный токен)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Обновить токены
      description: Принимает refresh token, возвращает новую пару access и refresh токенов (ротация).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        200:
          description: Новая пара токенов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        400:
          description: refreshToken не указан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Неверный или истекший refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # --- ТОВАРЫ ---
  /products:
    get:
      tags:
        - Products
      summary: Получить список всех товаров
      description: Возвращает массив всех товаров. Требует авторизации.
      security:
        - bearerAuth: []
      responses:
        200:
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        401:
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Products
      summary: Создать новый товар
      description: Добавляет новый товар в каталог. Авторизация не требуется (согласно вашему коду).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        201:
          description: Товар успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        400:
          description: Не все обязательные поля заполнены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /products/{id}:
    get:
      tags:
        - Products
      summary: Получить товар по ID
      description: Возвращает данные конкретного товара. Авторизация не требуется.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID товара
      responses:
        200:
          description: Данные товара
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        404:
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags:
        - Products
      summary: Полное обновление товара
      description: Заменяет все поля товара новыми данными. Требует авторизации.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID товара
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        200:
          description: Товар успешно обновлен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        400:
          description: Не все обязательные поля заполнены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - Products
      summary: Удалить товар
      description: Удаляет товар из каталога. Требует авторизации.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID товара
      responses:
        204:
          description: Товар успешно удален (нет содержимого в ответе)
        401:
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"